Fix for bug #23

+++ model/queue.php
@@ -719,6 +719,13 @@ function change_queue_state($course_id, $state){
     mysqli_close($sql_conn);
     return $entry["state"];
   }elseif($state == "closed"){ //By deleting the entry in queue_state, we cascade the other entries
+    //The above comment is correct, but cascading the other entries DOES NOT execute the triggers
+    //We need to manually delete the TAs still on duty so that the delete trigger will fire.
+    $query = "DELETE FROM ta_status WHERE course_id = '".$course_id."'";
+    if(!mysqli_query($sql_conn, $query)){
+      mysqli_close($sql_conn);
+      return -1;
+    }
     $query = "DELETE FROM queue_state WHERE course_id = '".$course_id."'";
   }elseif($state == 'frozen' || $state == 'open'){ //Since REPLACE calls DELETE then INSERT, calling REPLACE would CASCADE all other tables, we use ON DUPLICATE KEY UPDATE instead
     //Check if the course is disabled
